Implemented FastProcess
PlayerChallenges and Profiles now use the newly implement FastProcess
and FastProcessThreadSafe. These functions utilize LitJson. This results
in much faster deserialization of data using a fraction of the memory.
This comes at the cost of flexibility as full classes must exist for
this deserialization method to work. For other Playtomic data types
PJSON is more appropriate due to its flexibility